+++ common/src/main/java/com/telepathicgrunt/the_bumblezone/mixin/world/ServerWorldMixin.java
@@ -31,7 +31,7 @@ public abstract class ServerWorldMixin extends Level {
 	@Unique
 	private static int thebumblezone_counter = 0;
 
-
+	//TODO: use events
 	@Inject(
 			method = "tick(Ljava/util/function/BooleanSupplier;)V",
 			at = @At(value = "HEAD")