+++ common/src/main/java/com/telepathicgrunt/the_bumblezone/modinit/BzCriterias.java
@@ -61,6 +61,7 @@ public class BzCriterias {
     public static final CounterTrigger HONEY_SLIME_BRED_TRIGGER = new CounterTrigger(new ResourceLocation(Bumblezone.MODID, "honey_slime_bred"));
     public static final GenericTrigger HONEY_SLIME_CREATION_TRIGGER = new GenericTrigger(new ResourceLocation(Bumblezone.MODID, "honey_slime_creation"));
     public static final GenericTrigger HONEY_SLIME_HARVEST_TRIGGER = new GenericTrigger(new ResourceLocation(Bumblezone.MODID, "honey_slime_harvest"));
+    public static final GenericTrigger IS_NEAR_BEEHIVE_TRIGGER = new GenericTrigger(new ResourceLocation(Bumblezone.MODID, "is_near_beehive"));
     public static final KilledCounterTrigger KILLED_COUNTER_TRIGGER = new KilledCounterTrigger(new ResourceLocation(Bumblezone.MODID, "killed_counter"));
     public static final GenericTrigger LIGHT_SOUL_INCENSE_CANDLE_TRIGGER = new GenericTrigger(new ResourceLocation(Bumblezone.MODID, "light_soul_incense_candle"));
     public static final GenericTrigger POLLEN_PUFF_FIREBALL_TRIGGER = new GenericTrigger(new ResourceLocation(Bumblezone.MODID, "pollen_puff_fireball"));
@@ -130,6 +131,7 @@ public class BzCriterias {
         register(HONEY_SLIME_BRED_TRIGGER);
         register(HONEY_SLIME_CREATION_TRIGGER);
         register(HONEY_SLIME_HARVEST_TRIGGER);
+        register(IS_NEAR_BEEHIVE_TRIGGER);
         register(KILLED_COUNTER_TRIGGER);
         register(LIGHT_SOUL_INCENSE_CANDLE_TRIGGER);
         register(POLLEN_PUFF_MOOSHROOM_TRIGGER);